ai_backend/detection_pipeline/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
motor==3.4.0
pymongo==4.7.2
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9
apscheduler==3.10.4
opencv-python>=4.10.0.84
mediapipe>=0.10.14
numpy>=1.26.4
onnxruntime>=1.18.0
httpx>=0.27.0
requests>=2.31.0